# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.AgentRegistrationRequest:
    properties:
      agent_id:
        type: string
      capabilities:
        additionalProperties: true
        type: object
      description:
        type: string
      endpoint:
        type: string
      labels:
        items:
          type: string
        type: array
      name:
        type: string
      token:
        type: string
    required:
    - agent_id
    - name
    - token
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@iac-platform.com
    name: API Support
  description: Infrastructure as Code Platform API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: IAC Platform API
  version: "1.0"
paths:
...
  /api/v1/agent-pools:
    get:
      consumes:
      - application/json
      description: 获取所有Agent Pool
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回Pool列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取Agent Pool列表
      tags:
      - Agent Pool
    post:
      consumes:
      - application/json
      description: 创建新的Agent Pool
      parameters:
      - description: Pool信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数无效
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 创建Agent Pool
      tags:
      - Agent Pool
  /api/v1/agent-pools/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的Agent Pool
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的Pool ID
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 删除Agent Pool
      tags:
      - Agent Pool
    get:
      consumes:
      - application/json
      description: 根据ID获取Agent Pool详情
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回Pool详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的Pool ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Pool不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取Agent Pool详情
      tags:
      - Agent Pool
    put:
      consumes:
      - application/json
      description: 更新Agent Pool配置
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数无效
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新Agent Pool
      tags:
      - Agent Pool
  /api/v1/agent-pools/{id}/agents:
    post:
      consumes:
      - application/json
      description: 将Agent添加到指定的Pool
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: integer
      - description: Agent ID
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数无效
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 添加Agent到Pool
      tags:
      - Agent Pool
  /api/v1/agent-pools/{id}/agents/{agent_id}:
    delete:
      consumes:
      - application/json
      description: 从指定Pool中移除Agent
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agent_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的参数
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 从Pool移除Agent
      tags:
      - Agent Pool
  /api/v1/agents:
    get:
      consumes:
      - application/json
      description: 获取Agent列表，支持按状态和标签过滤
      parameters:
      - description: 状态过滤
        in: query
        name: status
        type: string
      - description: 标签过滤（逗号分隔）
        in: query
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回Agent列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取Agent列表
      tags:
      - Agent
  /api/v1/agents/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的Agent
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的Agent ID
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 删除Agent
      tags:
      - Agent
    get:
      consumes:
      - application/json
      description: 根据ID获取Agent的详细信息
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回Agent详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的Agent ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Agent不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取Agent详情
      tags:
      - Agent
    put:
      consumes:
      - application/json
      description: 更新Agent的配置信息
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数无效
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新Agent信息
      tags:
      - Agent
...
  /api/v1/agents/register:
    post:
      consumes:
      - application/json
      description: 注册新的Agent到平台
      parameters:
      - description: Agent注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AgentRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数无效
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 注册失败
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 注册Agent
      tags:
      - Agent
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"